requirements.txt
imageio==2.6.1
matplotlib==2.2.2
numpy==1.17.0
Pillow==5.1.0
scipy==1.0.1
tensorflow-gpu==1.14.0
